Refactor services and remove code duplication
