+++ backend/app/Http/Controllers/ScheduleSlotsController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\ScheduleSlots;
 use App\Models\User;
 use Illuminate\Http\Request;
-use App\Utils\Logger;
+use App\Services\Logger;
 
 class ScheduleSlotsController extends Controller
 {
