+++ backend/app/Http/Controllers/TrainingCourseTypeController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\TrainingCourseType;
 use App\Models\User;
 use Illuminate\Http\Request;
-use App\Utils\Logger;
+use App\Services\Logger;
 
 class TrainingCourseTypeController extends Controller
 {
