+++ backend/app/Services/DBTricks.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Utils;
+namespace App\Services;
 
 use Illuminate\Support\Facades\DB;
