+++ backend/app/Services/Logger.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Utils;
+namespace App\Services;
 
 use App\Models\User;
 use App\Models\Log;
