+++ backend/app/Services/TelegramBot.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Utils;
+namespace App\Services;
 
 use App\Models\TelegramBotNotifications;
 use App\Models\TelegramBotLogins;
